requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
alembic
pydantic
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
python-multipart
httpx
exa-py
firebase-admin
langchain-google-genai
pydantic[email]
langchain_core
langchain
pymupdf
python-docx
requests
